Refactor RigL callback into a mixin.
