+++ projects/transformers/callbacks/sparsity.py
@@ -29,11 +29,6 @@ from pandas import DataFrame
 from torch import count_nonzero
 from transformers import TrainerCallback
 
-from nupic.research.frameworks.dynamic_sparse import (
-    CosineDecayPruneScheduler,
-    global_add_by_abs_grad,
-    global_prune_by_abs_weight,
-)
 from nupic.research.frameworks.pytorch.model_utils import (
     count_nonzero_params,
     filter_modules,
@@ -42,7 +37,6 @@ from nupic.torch.modules.sparse_weights import SparseWeightsBase, rezero_weights
 
 __all__ = [
     "RezeroWeightsCallback",
-    "RigLCallback",
     "PlotDensitiesCallback"
 ]
 
@@ -93,108 +87,6 @@ class RezeroWeightsCallback(TrainerCallback):
                 encoder_sparsity=encoder_sparsity
             )
 
-            wandb.log(logs, step=state.global_step)
-
-
-class RigLCallback(TrainerCallback):
-    """
-    Perform `RigL`_ dynamic sparsity every N steps on the sparse modules within the
-    network. Weights are pruned according to their ranks absolute values and regrowing
-    (adding weights) occurs by ranking their absolute gradients. As per the original
-    paper, the pruning decreases over time according to a cosine decay.
-
-    .. _RigL: https://arxiv.org/abs/1911.11134
-
-    :param prune_fraction: initial fraction of weights to prune
-    :param prune_freq: how often, in iterations, to prune and regrow weights
-    :param warmup_steps: defaults to prune_freq; e.g. prune_freq of 100 will allow 100
-                         steps before pruning for the first time
-    """
-
-    def __init__(
-        self,
-        prune_fraction=0.3,
-        prune_freq=100,
-        warmup_steps=None
-    ):
-        self.prune_fraction = prune_fraction
-        self.prune_freq = prune_freq
-        self.prune_scheduler = None
-        self.warmup_steps = warmup_steps
-
-    def on_init_end(self, args, state, control, model, optimizer=None, **kwargs):
-        """Save a list of the sparse modules and initialize the pruning schedule"""
-
-        warmup_steps = self.warmup_steps or self.prune_freq
-        self.prune_scheduler = CosineDecayPruneScheduler(
-            total_steps=args.max_steps,
-            prune_fraction=self.prune_fraction,
-            warmup_steps=warmup_steps - 1
-        )
-        self.sparse_modules = filter_modules(
-            model, include_modules=[SparseWeightsBase]
-        ).values()
-
-    def on_step_end(
-        self, args, state, control, model=None,
-        train_dataloader=None, optimizer=None, **kwargs
-    ):
-        """Prune and regrow weights every 'prune_freq' iterations."""
-
-        if state.global_step % self.prune_freq != 0:
-            self.prune_scheduler.step()
-            return
-
-        # Pre-prune sparsities.
-        param_sparsity0, mask_sparsity0 = calc_cumulative_sparsity(self.sparse_modules)
-
-        # Prune weights.
-        model.apply(rezero_weights)
-        prune_fraction = self.prune_scheduler.get_prune_fraction()
-        num_removed = global_prune_by_abs_weight(self.sparse_modules, prune_fraction)
-        model.apply(rezero_weights)
-
-        # Post-prune sparsities.
-        param_sparsity1, mask_sparsity1 = calc_cumulative_sparsity(self.sparse_modules)
-
-        # Accumulate gradients over one batch.
-        optimizer.zero_grad()
-        train_batch = next(iter(train_dataloader))
-        inputs_to_device(train_batch, device=args.device)
-        output = model(**train_batch)
-        output.loss.backward()
-
-        # Regrow weights
-        num_add = self.prune_scheduler.get_num_add(num_removed)
-        global_add_by_abs_grad(self.sparse_modules, num_add)
-        self.prune_scheduler.step()
-
-        # Post-grow sparsities.
-        param_sparsity2, mask_sparsity2 = calc_cumulative_sparsity(self.sparse_modules)
-
-        # Log pruning stats.
-        actual_pruned = param_sparsity1 - param_sparsity0
-        actual_pruned_on_params = actual_pruned / (1 - mask_sparsity0)
-
-        logging.info(f"RigLCallback:")
-        logging.info(f"Target: remove {prune_fraction} frac of on params")
-        logging.info(f"Actual: removed {actual_pruned_on_params} fraction of on params")
-
-        # For now, the logs are very robust to ensure pruning occurs as expected.
-        # TODO: Remove non-essential logging.
-        logs = dict({
-            "rigl/target_pruned_on_params": prune_fraction,
-            "rigl/actual_pruned_on_params": actual_pruned_on_params,
-            "rigl/target_pruned_all_params": prune_fraction * mask_sparsity0,
-            "rigl/actual_pruned_all_params": actual_pruned,
-            "rigl/pre_prune_param_sparsity": param_sparsity0,
-            "rigl/pre_prune_mask_sparsity": mask_sparsity0,
-            "rigl/post_prune_param_sparsity": param_sparsity1,
-            "rigl/post_prune_mask_sparsity": mask_sparsity1,
-            "rigl/pre_grow_param_sparsity": param_sparsity2,
-            "rigl/post_grow_mask_sparsity": mask_sparsity2,
-        })
-        if wandb.run is not None:
             wandb.log(logs, step=state.global_step)
 
 
@@ -289,14 +181,6 @@ class PlotDensitiesCallback(TrainerCallback):
 # Utilities
 # -------------
 
-def inputs_to_device(inputs, device):
-    """
-    Prep inputs dict by transferring all inputs to the device.
-    """
-    for k, v in inputs.items():
-        if isinstance(v, torch.Tensor):
-            inputs[k] = v.to(device)
-
 
 def calc_sparsity(tensor):
     """Calculate the sparsity of a given tensor."""
@@ -310,24 +194,6 @@ def calc_model_sparsity(model):
     tot, nz = count_nonzero_params(model)
     sparsity = 1 - nz / tot
     return sparsity
-
-
-def calc_cumulative_sparsity(sparse_modules):
-    """
-    Calculate the sparsities across a list of sparse modules. Both the weight sparsity
-    and the zero mask sparsity is calculated.
-    """
-    total_off = 0
-    total_zero = 0
-    total_params = 0
-    for m in sparse_modules:
-        total_off += count_nonzero(m.zero_mask)
-        total_zero += m.weight.numel() - count_nonzero(m.weight)
-        total_params += m.weight.numel()
-
-    mask_sparsity = total_off / total_params
-    param_sparsity = total_zero / total_params
-    return param_sparsity, mask_sparsity
 
 
 def get_density_by_layer(sparse_modules):
